feat(install-dependencies): adds install-dependencies fn
